feat(store): add categories in store
