+++ src/components/Cashier/TabScroller.vue
@@ -1,38 +1,29 @@
 <script setup lang='ts'>
 import { ref } from 'vue'
+import { useProducts } from '../../stores/products';
 
-const categories = [
-  [ 'Favorite', 'fad fa-stars' ],
-  [ 'Foods', 'fad fa-utensils' ],
-  [ 'Cold drinks', 'fad fa-beer' ],
-  [ 'Hot drinks', 'fad fa-coffee' ],
-  [ 'Search', 'fad fa-search' ],
-  [ 'Repeat', 'fad fa-cogs' ],
-  [ 'Foods', 'fad fa-utensils' ],
-  [ 'Cold drinks', 'fad fa-beer' ],
-  [ 'Hot drinks', 'fad fa-coffee' ],
-  [ 'Search', 'fad fa-search' ],
-  [ 'Repeat', 'fad fa-cogs' ],
-];
+interface ScrollTo {
+  (element: HTMLDivElement, step: number, duration: number): void
+}
+// type ScrollTo= (element: HTMLDivElement, step: number, duration: number) => void
 
-const tabActive = ref(1)
-const wrapper = ref<HTMLDivElement | null>(null)
+const products = useProducts();
+
+const categories = ref(products.categories);
+const tabActive = ref(products.tabActive);
+// const wrapper = ref<HTMLDivElement | null>(null)
+const wrapper = ref()
 
 const step = 200
-const duration = 200
-let position = 0
+const duration = 100
 const tabGoLeft = () => {
-  console.log('we go left!')
-  scrollTo(wrapper.value as HTMLDivElement, -step, duration)
+  scrollTo(wrapper.value, -step, duration)
 }
 const tabGoRight = () => {
-  console.log('we go right!')
   scrollTo(wrapper.value as HTMLDivElement, step, duration)
 }
-const scrollTo = (element: HTMLDivElement, step: number, duration:number) : void => {
+const scrollTo: ScrollTo = (element, step, duration) => {
   const scrollPos = element.scrollLeft
-  position = element.scrollLeft
-  console.log('lets scroll', scrollPos, position)
 
   // condition to check if scrolling is required
   if (!((scrollPos === 0 || step > 0)
